requirements.txt
beautifulsoup4==4.8.0
blinker==1.4
cchardet==2.1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
html5lib==1.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
webencodings==0.5.1
Werkzeug==0.15.5